+++ src/integrationTest/java/com/pubnub/api/integration/objects/memberships/CustomMetadataInMembershipPropagationIT.java
@@ -1,21 +1,29 @@
 package com.pubnub.api.integration.objects.memberships;
 
+import com.pubnub.api.PubNub;
 import com.pubnub.api.PubNubException;
+import com.pubnub.api.integration.managers.subscription.SubscribeCallbackAdapter;
 import com.pubnub.api.integration.objects.ObjectsApiBaseIT;
 import com.pubnub.api.models.consumer.objects_api.channel.PNSetChannelMetadataResult;
 import com.pubnub.api.models.consumer.objects_api.membership.PNChannelMembership;
 import com.pubnub.api.models.consumer.objects_api.membership.PNGetMembershipsResult;
+import com.pubnub.api.models.consumer.objects_api.membership.PNMembershipResult;
 import com.pubnub.api.models.consumer.objects_api.membership.PNSetMembershipResult;
+import org.awaitility.core.ThrowingRunnable;
 import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
 
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.UUID;
+import java.util.concurrent.CopyOnWriteArrayList;
+import java.util.concurrent.TimeUnit;
 
 import static com.pubnub.api.endpoints.objects_api.utils.Include.PNChannelDetailsLevel.CHANNEL;
 import static com.pubnub.api.endpoints.objects_api.utils.Include.PNChannelDetailsLevel.CHANNEL_WITH_CUSTOM;
+import static org.awaitility.Awaitility.await;
 import static org.hamcrest.Matchers.allOf;
 import static org.hamcrest.Matchers.hasItem;
 import static org.hamcrest.Matchers.hasProperty;
@@ -38,6 +46,22 @@ public class CustomMetadataInMembershipPropagationIT extends ObjectsApiBaseIT {
     private PNSetChannelMetadataResult setChannelMetadataResult;
     private PNSetMembershipResult setMembershipResult;
 
+
+    private CopyOnWriteArrayList<PNMembershipResult> pnMembershipResults = new CopyOnWriteArrayList<>();
+
+    @Before
+    public void setCallbackListener() {
+        pubNubUnderTest.addListener(new SubscribeCallbackAdapter() {
+            @Override
+            public void membership(final PubNub pubnub, final PNMembershipResult pnMembershipResult) {
+                pnMembershipResults.add(pnMembershipResult);
+            }
+        });
+
+        pubNubUnderTest.subscribe()
+                .channels(Collections.singletonList(testChannelMetadataId))
+                .execute();
+    }
 
     @Test
     public void setMembershipCustomHappyPath() throws PubNubException {
@@ -86,6 +110,14 @@ public class CustomMetadataInMembershipPropagationIT extends ObjectsApiBaseIT {
                                                 hasProperty("name", is(testChannelName)),
                                                 hasProperty("description", is(testDescription)),
                                                 hasProperty("custom", nullValue()))))))));
+
+        await().atMost(1, TimeUnit.SECONDS).untilAsserted(new ThrowingRunnable() {
+            @Override
+            public void run() throws Throwable {
+                assertThat(pnMembershipResults, hasItem(
+                        hasProperty("data", hasProperty("uuid", is(pubNubUnderTest.getConfiguration().getUuid())))));
+            }
+        });
     }
 
     @After