+++ src/main/java/com/pubnub/api/models/consumer/objects_api/membership/PNMembership.java
@@ -3,13 +3,12 @@ package com.pubnub.api.models.consumer.objects_api.membership;
 import com.google.gson.annotations.JsonAdapter;
 import com.pubnub.api.models.consumer.objects_api.channel.PNChannelMetadata;
 import com.pubnub.api.models.consumer.objects_api.util.CustomPayloadJsonInterceptor;
+import com.pubnub.api.utils.UnwrapSingleField;
 import lombok.*;
 import lombok.experimental.Accessors;
 
-@Getter
+@Data
 @Accessors(chain = true)
-@EqualsAndHashCode
-@ToString
 @RequiredArgsConstructor
 public class PNMembership {
     @NonNull
@@ -17,6 +16,9 @@ public class PNMembership {
 
     @JsonAdapter(CustomPayloadJsonInterceptor.class)
     protected Object custom;
+
+    @JsonAdapter(UnwrapSingleField.class)
+    protected String uuid;
 
     protected String updated;
     protected String eTag;