+++ src/main/java/com/pubnub/api/models/consumer/pubsub/objects/ObjectResult.java
@@ -9,6 +9,7 @@ public abstract class ObjectResult<T> extends BasePubSubResult {
 
     @Getter
     protected String event;
+    @Getter
     protected T data;
 
     public ObjectResult(BasePubSubResult result, String event, T data) {